re-styling for Landing page and re-preparing file structure for landingpage
