+++ client/src/App/style.css
@@ -1,4 +1,4 @@
-body {
+/* body {
   font-size: 16px;
   line-height: 1.6;
   overflow-x: hidden;
@@ -127,7 +127,6 @@ p {
   opacity: 1;
 }
 
-/* ant menu */
 .header .ant-menu {
   border: 0;
   background: none !important;
@@ -184,10 +183,7 @@ p {
   opacity: 1;
 }
 
-/* 
-hero
----- 
-*/
+
 .heroBlock {
   background: url("../assets/images/header.png") no-repeat;
   background-position: 50% 100%;
@@ -246,10 +242,7 @@ hero
   margin: 0 10px 0 0;
 }
 
-/* 
-footer
-------
-*/
+
 .ant-layout-footer {
   background: #546681;
   color: #fff;
@@ -285,10 +278,7 @@ span.icon {
   display: block;
 }
 
-/* 
-responsive
-----------
-*/
+
 @media only screen and (max-width: 991px) {
   .pricingBlock .ant-row > div:nth-child(2) .ant-list-item {
     transform: scale(1.14);
@@ -333,4 +323,4 @@ responsive
 .formItem {
   font-size: "large";
   margin: 30px 0 30px 0;
-}
+} */
