+++ client/src/Layout/LandingPage/index.tsx
@@ -1,0 +1,4 @@
+import HeaderSection from './header';
+import Footer from './footer';
+
+export { HeaderSection, Footer };
