--- client/src/components/FlatContainer/style.css
@@ -1,3 +1,0 @@
-#ads {
-  background-color: rgb(246, 248, 250) !important
-}
