+++ client/src/components/LandingPage/HomeSection.tsx
@@ -1,7 +1,11 @@
-import { Carousel, Button } from 'antd';
-import staticData from '../../StaticData';
+import { Carousel, Anchor } from 'antd';
 
-const AppHome: React.FC = () => (
+import staticData from '../../StaticData';
+import './style.css';
+
+const { Link } = Anchor;
+
+const HomeSection: React.FC = () => (
   <div id="home" className="heroBlock">
     <div className="overlay">
       <Carousel>
@@ -11,13 +15,15 @@ const AppHome: React.FC = () => (
               <h3>{item.title}</h3>
               <p>{item.content}</p>
               <div className="btnHolder">
-                <Button type="primary" size="large" id="about-us">المزيد عنا</Button>
-                <a href="#video">
-                  <Button size="large" id="video">
+                <Anchor targetOffset={65} className="btnHolder">
+                  <Link key="about-us" href="#about-us" title="المزيد عنا">
                     <i className="ri-slideshow-3-line" />
-                    شاهدنا
-                  </Button>
-                </a>
+                  </Link>
+                  <Link key="video" href="#video" title="شاهدنا">
+                    <i className="ri-slideshow-3-line" />
+                  </Link>
+                </Anchor>
+
               </div>
             </div>
           </div>
@@ -27,4 +33,4 @@ const AppHome: React.FC = () => (
   </div>
 );
 
-export default AppHome;
+export default HomeSection;
