+++ client/src/components/LandingPage/VideoSection.tsx
@@ -1,0 +1,8 @@
+import ReactPlayer from 'react-player';
+import './style.css';
+
+const VideoSection: React.FC = () => (
+  <ReactPlayer id="video" url="https://www.youtube.com/watch?v=ysz5S6PUM-U" />
+);
+
+export default VideoSection;
