+++ client/src/components/LandingPage/index.tsx
@@ -1,0 +1,12 @@
+/* eslint-disable import/no-cycle */
+import HomeSection from './HomeSection';
+import AboutUs from './AboutUs';
+import AdsContainer from './AdsContainer';
+import FlatContainer from './FlatContainer';
+import Services from './Services';
+import VideoSection from './VideoSection';
+import ContactUs from './ContactUs';
+
+export {
+  HomeSection, AboutUs, AdsContainer, FlatContainer, Services, VideoSection, ContactUs,
+};
