--- client/src/components/VideoSection/style.css
@@ -1,38 +1,0 @@
-.img-about {
-    width: 460px;
-    height: auto;
-    border-radius: 8px;
-}
-
-.ant-col.ant-col-12.ant-col-rtl {
-    display: flex;
-    justify-content: space-around;
-}
-
-#contact-section {
-  align-items: center !important;
-}
-.video {
-    width: 100% !important;
-}
-form#nest-messages {
-    width: 100%;
-    margin-left: 16px;
-}
-.ant-row.ant-row-rtl.ant-form-item-row {
-    display: flex;
-    flex-direction: column;
-    align-items: flex-start;
-}
-.ant-form-horizontal .ant-form-item-control {
-    width: 100%;
-}
-textarea#nest-messages_description {
-    height: 140px;
-}
-
-
-#contact-section .ant-col.ant-col-rtl.ant-col-sm-24.ant-col-sm-offset-0.ant-col-xl-12.ant-col-xl-offset-0 {
-    display: flex;
-    justify-content: flex-end;
-}
