+++ server/database/config/fakeData.ts
@@ -557,21 +557,21 @@ const flats = [
   {
     flat_number: 101,
     area: 180,
-    notes: 'شقة غربية',
+    notes: 'شقة أرضية بمواصفات مميزة جدا · شقة جاهزة للبيع شرق برشلونة · شقة للببع خلف شارع الصناعة',
     is_active: true,
     UserId: 1,
   },
   {
     flat_number: 102,
     area: 160,
-    notes: 'شقة جنوبية',
+    notes: 'شقة جنوبية بمواصفات مميزة جدا · شقة جاهزة للبيع شرق برشلونة · شقة للببع خلف شارع الصناعة',
     is_active: true,
     UserId: 2,
   },
   {
     flat_number: 103,
     area: 120,
-    notes: 'شقة شرقية',
+    notes: 'شقة أرضية بمواصفات مميزة جدا · شقة جاهزة للبيع شرق برشلونة · شقة للببع خلف شارع الصناعة',
     is_active: true,
     UserId: 3,
   },
